+++ src/gigachanie/context/__init__.py
@@ -2,13 +2,15 @@
 
 from gigachanie.context.memory import MemoryEntry, MemoryStore
 from gigachanie.context.project_file import ProjectContext, load_project_context
-from gigachanie.context.refs import expand_file_refs
+from gigachanie.context.refs import ExpandedRefs, expand_file_refs, expand_refs
 from gigachanie.context.repo_map import RepoMap, build_repo_map
 
 __all__ = [
     "ProjectContext",
     "load_project_context",
     "expand_file_refs",
+    "expand_refs",
+    "ExpandedRefs",
     "RepoMap",
     "build_repo_map",
     "MemoryStore",